feat: move UI creator components with now-resolved deps to commons
Move the following files to commons KMP module:
- IZapRaiser interface
- ZapRaiserRequest composable
- ForwardZapToButton + ZapSplitIcon composables
- ContentSensitivityExplainer composable
- SecretEmojiRequest composable

Also adds:
- Required string resources to commons strings.xml
- Shape.kt theme helpers (QuoteBorder, Size20Modifier, DividerThickness, placeholderText)

Original amethyst files now delegate to commons implementations.
Build verified: commons:compileKotlinJvm + amethyst:compilePlayDebugKotlin pass.
