+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/note/creators/secretEmoji/SecretEmojiRequest.kt
@@ -20,115 +20,10 @@
  */
 package com.vitorpamplona.amethyst.ui.note.creators.secretEmoji
 
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Assistant
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.KeyboardCapitalization
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.vitorpamplona.amethyst.R
-import com.vitorpamplona.amethyst.commons.emojicoder.EmojiCoder
-import com.vitorpamplona.amethyst.ui.stringRes
-import com.vitorpamplona.amethyst.ui.theme.DividerThickness
-import com.vitorpamplona.amethyst.ui.theme.QuoteBorder
-import com.vitorpamplona.amethyst.ui.theme.Size20Modifier
-import com.vitorpamplona.amethyst.ui.theme.placeholderText
+import com.vitorpamplona.amethyst.commons.ui.note.creators.secretEmoji.SecretEmojiRequest as CommonsSecretEmojiRequest
 
 @Composable
 fun SecretEmojiRequest(onSuccess: (String) -> Unit) {
-    Column {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier.fillMaxWidth().padding(bottom = 5.dp),
-        ) {
-            Icon(
-                imageVector = Icons.Outlined.Assistant,
-                null,
-                modifier = Size20Modifier,
-                tint = MaterialTheme.colorScheme.onBackground,
-            )
-
-            Text(
-                text = stringRes(R.string.secret_emoji_maker),
-                fontSize = 20.sp,
-                fontWeight = FontWeight.W500,
-                modifier = Modifier.padding(start = 10.dp),
-            )
-        }
-
-        HorizontalDivider(thickness = DividerThickness)
-
-        var secretMessage by remember { mutableStateOf("") }
-        var publicPrefix by remember { mutableStateOf("") }
-
-        OutlinedTextField(
-            label = { Text(text = stringRes(R.string.secret_note_to_receiver)) },
-            modifier = Modifier.fillMaxWidth(),
-            value = secretMessage,
-            onValueChange = { secretMessage = it },
-            placeholder = {
-                Text(
-                    text = stringRes(R.string.secret_note_to_receiver_placeholder),
-                    color = MaterialTheme.colorScheme.placeholderText,
-                )
-            },
-            keyboardOptions =
-                KeyboardOptions.Default.copy(
-                    capitalization = KeyboardCapitalization.Sentences,
-                ),
-            singleLine = true,
-        )
-
-        OutlinedTextField(
-            label = { Text(text = stringRes(R.string.secret_visible_text)) },
-            modifier = Modifier.fillMaxWidth(),
-            value = publicPrefix,
-            onValueChange = { publicPrefix = it },
-            placeholder = {
-                Text(
-                    text = stringRes(R.string.secret_visible_text_placeholder),
-                    color = MaterialTheme.colorScheme.placeholderText,
-                )
-            },
-            singleLine = true,
-        )
-
-        Button(
-            modifier = Modifier.fillMaxWidth().padding(vertical = 10.dp),
-            onClick = {
-                onSuccess(EmojiCoder.encode(publicPrefix, secretMessage))
-            },
-            shape = QuoteBorder,
-            colors =
-                ButtonDefaults.buttonColors(
-                    containerColor = MaterialTheme.colorScheme.primary,
-                ),
-        ) {
-            Text(
-                text = stringRes(R.string.secret_add_to_text),
-                color = Color.White,
-                fontSize = 20.sp,
-            )
-        }
-    }
+    CommonsSecretEmojiRequest(onSuccess = onSuccess)
 }
