+++ commons/src/commonMain/composeResources/values/strings.xml
@@ -54,4 +54,27 @@
     <!-- Accessibility -->
     <string name="accessibility_user_avatar">User avatar</string>
     <string name="accessibility_navigate">Navigate</string>
+
+    <!-- Zapraiser -->
+    <string name="zapraiser">Zapraiser</string>
+    <string name="zapraiser_explainer">Adds a target amount of sats to raise for this post. Supporting clients may show this as a progress bar to incentivize donations</string>
+    <string name="zapraiser_target_amount_in_sats">Target Amount in Sats</string>
+
+    <!-- Zap Split -->
+    <string name="zap_split_title">Split and Forward Zaps</string>
+    <string name="zaps">Zaps</string>
+
+    <!-- Content Warning -->
+    <string name="content_warning">Content warning</string>
+    <string name="add_sensitive_content_label">Sensitive Content</string>
+    <string name="add_sensitive_content_explainer">Adds sensitive content warning before showing your content. This is ideal for any NSFW content or content some people may find offensive or disturbing</string>
+    <string name="add_sensitive_content_description_placeholder">Reason (optional)</string>
+
+    <!-- Secret Emoji -->
+    <string name="secret_emoji_maker">Secret Emoji Maker</string>
+    <string name="secret_note_to_receiver">Secret Note to Receiver</string>
+    <string name="secret_note_to_receiver_placeholder">My hidden message</string>
+    <string name="secret_visible_text">Visible Prefix</string>
+    <string name="secret_visible_text_placeholder">😎</string>
+    <string name="secret_add_to_text">Add to Post</string>
 </resources>
